requirements.txt
# FastAPI URL Shortener
# Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
validators==0.22.0